Move `FixedParameters` to `pyFAI.containers`
